backend/requirements.txt
Django==3.2.25
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
Pillow==10.2.0
django-filter==23.5
